Moved unit tests for rationalise_blend_time_coords into plugin test file.
